+++ unicode-data-scripts/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.1 TBD
+
+* Add `unicodeVersion` to `Unicode.Char.General.Scripts`.
+
 ## 0.2.0.1 (December 2022)
 
 - Fix [Unicode scripts handling on big-endian architectures](https://github.com/composewell/unicode-data/issues/97).