+++ unicode-data-security/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.1 TBD
+
+- Add `unicodeVersion` to `Unicode.Char.Identifiers.Security`.
+
 ## 0.2.0 (September 2022)
 
 - Update to [Unicode 15.0.0](https://www.unicode.org/versions/Unicode15.0.0/).